refactor: created addWordsToTextarea. Closes #8
